+++ handlers/zuora-salesforce-link-remover/package.json
@@ -4,7 +4,7 @@
 	"scripts": {
 		"test": "jest --group=-integration",
 		"it-test": "jest --group=integration",
-		"build": "esbuild --bundle --platform=node --target=node20 --outdir=target src/**/*.ts",
+		"build": "esbuild --bundle --platform=node --target=node20 --outdir=target src/handlers/*.ts",
 		"lint": "eslint src/**/*.ts",
 		"package": "pnpm type-check && pnpm lint && pnpm test && pnpm build && cd target; zip -qr zuora-salesforce-link-remover.zip ./*.js",
 		"type-check": "tsc --noEmit",
