+++ handlers/zuora-salesforce-link-remover/src/handlers/updateSfBillingAccounts.ts
@@ -1,15 +1,15 @@
 import { getSecretValue } from '@modules/secrets-manager/src/getSecret';
 import { stageFromEnvironment } from '@modules/stage';
 import type { Handler } from 'aws-lambda';
-import { BillingAccountRecordsSchema, doSfAuth, updateSfBillingAccounts } from './salesforceHttp';
+import { BillingAccountRecordsSchema, doSfAuth, updateSfBillingAccounts } from '../salesforceHttp';
 import type {
 	BillingAccountRecord,
 	SalesforceUpdateResponseArray,
 	SfApiUserAuth,
 	SfConnectedAppAuth,
-} from './salesforceHttp';
-import { getSalesforceSecretNames } from './secrets';
-import type { ApiUserSecret, ConnectedAppSecret } from './secrets';
+} from '../salesforceHttp';
+import { getSalesforceSecretNames } from '../secrets';
+import type { ApiUserSecret, ConnectedAppSecret } from '../secrets';
 
 export const handler: Handler<BillingAccountRecord[], SalesforceUpdateResponseArray> = async (billingAccounts) => {
